Change tile / token mapping
The tile to / from token mapping is of bigger importance, and the grid must be aware of this mapping.  The tiles now include a token id, and the grid includes a mapping between the verticies on the grid and the token id.  This allows for better whole-tile highlighting as well as future informational messages on tile selection.