Textures
Switched off of just colors per triangle to having a texture per triangle.

The selection and hover changed to replacing the triangles in the token with a different texture.

This led to adding a loading screen and texture map generator, and all the infrastructure required to support it.

There's still a lot of cleanup work to do after this big change, but it's a milestone.