Generate token texture modes
Rather than force each token type to include the hover/se/ect/etc modes, those are automatically generated off of the svg file.

This also alters the uv-map format and the generated image to be a bit nicer, and have a better 0-1 resolution to better take advantage of the horizintal space.