Fixed Grid
After experiments with multiple segments showed major performance issues with the naive implementation, the code is going through some drastic restructuring to make the graphics more static while allowing the game board to better reflect the state changes generated by interacting with the server API.

The server will store individual tiles by x/y coordinates, and it is up to the client to create an efficient sparse data store.  This reflects that change by breaking the game board state out of the redux store, because it has a separate memory management, and is expected to only interact with a single scene.

The new grid implementation will go back to the older drawing, because the object intersection was much faster and the drawing itself was fast.  This will be a more complicated data management problem, but it should be well containerized.  There's a lot of rethinking to do here.