Some rendering fixes
Fixed up some of the rendering so it shows up better.  Fixed up the server/client interaction a bit.

There's still a lot of work left for this setup.  The tile alignment visually is wrong (again), which seems to be throwing off a lot of other parts of the system.