Fix some bugs.
The main grid was being loaded several times over.  Fixed this partially by adding in a temporary loading screen for the various phases.

Added a loading texture.

Fixed a few places with incorrect tile loading.  There's still some massive bugs with the grid construction - the tokens are not properly mapped to a hex, which causes issues with tiles being null in the grid map pointers.  Tokens are still marked as loading when they aren't - these probably just need to be cleared to 'unset'.