A few fixes + Bump Map
Fixed up some of the map generation.  It was the test data generation that was creating bad tile alignments.

Added bump mapping.  This includes rewriting the texture generation.  It's now a javascript CLI that renders the images in parallel, which significantly increases execution time.

There's still bugs left with height calculations, not keeping the grid centered, and mouse over / selection.