Force the manual creation of the images.
The image files can be very big, and as a result don't belong in source control.  Instead, these are now built.  For the public site, that means building these as part of the public page generation.

This has the added benefit of making all image files forced to be re-creatable.  It means fonts need to be distributed with the source.