Refactor the grid storage
The grid in client memory now pre-calculates vertex height and adjacency, so that this doesn't have to be done every time the board is re-rendered.  This adds a bit more memory required for each tile, but has a big improvement on the performance.  Big improvements are on the tile height.

Unfortunately, tile height calculations have a bug right now that needs to be fixed.